+++ src/main.go
@@ -9,7 +9,7 @@ import (
 	tea "github.com/charmbracelet/bubbletea"
 )
 
-var version = "0.3.1"
+var version = "0.3.2"
 
 func main() {
 	var (